Move abstract types into a separate module
